% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betamix.R
\name{NegBetaLogLike}
\alias{NegBetaLogLike}
\title{Compute the (negative) log-likelihood}
\usage{
NegBetaLogLike(param, vec, z)
}
\arguments{
\item{param}{A vector of parameters of length 2K}

\item{vec}{A data vector of length N}

\item{z}{A matrix of size NxK of nonegative real numbers, with the sum of each row equal to 1.}
}
\description{
Compute the (negative) log-likelihood
}
